adds gameService methods

+++ src/main/java/com/example/Battleships_Backend/models/Grid.java
@@ -10,7 +10,7 @@ public class Grid {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
+    private int id;
 
     @Column(name = "player_name")
     private String playerName;
@@ -32,12 +32,11 @@ public class Grid {
 
     public Grid(){
     }
-
-    public Long getId() {
+    public int getId() {
         return id;
     }
 
-    public void setId(Long id) {
+    public void setId(int id) {
         this.id = id;
     }
